\name{collpcm.summaryplot}
\alias{collpcm.summaryplot}
\title{Make a summary plot of a \code{collpcm} run}
\description{
  \code{\link{collpcm.summaryplot}} creates a 2 by 2 summary plot showing traces from the MCMC run as well as the posterior KL positions for the most visited model.
}
\usage{
collpcm.summaryplot( x )
}
\arguments{
 \item{x}{An object of class \code{collpcm} }
}


\author{
Jason Wyse
}

\references{
 
 Ryan, C., Wyse, J. and Friel N. (2016)
 \emph{ Bayesian model selection for the latent position cluster model for Social Networks.} 
 Network Science (accepted).
 
}



