\name{Dolphins}
\alias{Dolphins}
\alias{Dolphins}
\docType{data}
\title{Dolphins}
\description{
Network describing  social ties between dolphins off doubtful sound.
}
\usage{
 data(Dolphins)
}
\source{
Lusseau, D., Schneider, K., Boisseau, O. J., Hasse, P., Slooten, E. and Dawson, S. M. (2003) \emph{The bottlenose dolphin community of Doubtful Sound features a large proportion of long-lasting associations- Can geographic isolation explain this unique trait?} Behavioural Ecology and Sociobiology 54, 396--405.
}

