\name{Karate}
\alias{Karate}
\alias{Karate}
\docType{data}
\title{Network describing loyalty in the Karate club.}
\description{
The well known Karate data.
}
\usage{
 data(Karate)
}
\source{
Zachary, W.~W. (1977) \emph{An information flow model for conflict and fission in small groups,} Journal of Anthropological Research 33, 452-473
}

