\name{monks}
\alias{monks}
\alias{monks}
\docType{data}
\title{Friendship network in Sampson's monastery}
\description{
The well known Monks data.
}
\usage{
 data(monks)
}
\source{
 Sampson, S.~F. (1968), \emph{A novitiate in a period of change: An
 experimental and case study of relationships,} 
 Unpublished Ph.D. dissertation, Department of Sociology, Cornell University.

 \url{http://vlado.fmf.uni-lj.si/pub/networks/data/esna/sampson.htm}
}

