\name{Monks}
\alias{Monks}
\alias{Monks}
\docType{data}
\title{Monks}
\description{
Sampson's aggregated Monk's dataset.
}
\usage{
 data(Monks)
}
\source{
 Sampson, S.~F. (1968), \emph{A novitiate in a period of change: An
 experimental and case study of relationships,} 
 Unpublished Ph.D. dissertation, Department of Sociology, Cornell University.

 \url{http://vlado.fmf.uni-lj.si/pub/networks/data/esna/sampson.htm}
}

