\name{print.collpcm}
\alias{print.collpcm}
\title{Print a collpcm object}
\description{
  Print a summary of a collpcm object.
}

\usage{
	\method{print}{collpcm}( x, \dots )
}

\arguments{
 \item{x}{An object of class \code{collpcm}.}
 \item{\dots}{Optional arguments to lower level functions.}
}

\author{  Jason Wyse }

\references{
 
  Ryan, C., Wyse, J. and Friel, N. (2017)
 \emph{Bayesian model selection for the latent position cluster model for Social Networks.} 
Network Science, volume 5, 70-91.
 
}



