\name{summary.collpcm}
\alias{summary.collpcm}
\title{Summarize a collpcm object}
\description{
  Print a summary of a collpcm object.
}

\usage{
	\method{summary}{collpcm}( object, \dots )
}

\arguments{
 \item{object}{An object of class \code{collpcm}.}
 \item{\dots}{Optional arguments to lower level functions.}
}

\author{  Jason Wyse }

\references{
 
 Ryan, C., Wyse, J. and Friel, N. (2017)
 \emph{Bayesian model selection for the latent position cluster model for Social Networks.} 
 Network Science, volume 5, 70-91.
 
}



